Fix posix diskusage

+++ diskusage/diskuage_posix.go
@@ -7,6 +7,7 @@ import (
 )
 
 // disk usage of path/disk
+//noinspection GoRedundantConversion
 func DiskUsage(path string) (*DiskStatus, error) {
 	fs := syscall.Statfs_t{}
 	err := syscall.Statfs(path, &fs)
@@ -14,8 +15,8 @@ func DiskUsage(path string) (*DiskStatus, error) {
 		return nil, err
 	}
 	status := &DiskStatus{
-		All:  int64(fs.Blocks) * fs.Bsize,
-		Free: int64(fs.Bfree) * fs.Bsize,
+		All:  int64(fs.Blocks) * int64(fs.Bsize),
+		Free: int64(fs.Bfree) * int64(fs.Bsize),
 	}
 	status.Used = status.All - status.Free
 	return status, nil
